Fix LIME explanation dtype mismatch error
Issue: When using LIME explanation method, the system threw:
"Input type (double) and bias type (float) should be the same"

Root cause: In the LIME predict_fn, tensors were being created without
explicit dtype specification. torch.from_numpy() was creating float64
(double) tensors when the model expects float32 (float).

Fix: Added explicit .float() calls to ensure all tensors are float32:
- Line 270: Added .float() after permute for individual tensors
- Line 273: Added .float() after stacking the batch tensor

This ensures consistent float32 dtype throughout the LIME pipeline,
matching what the models expect.

+++ crop_monitoring_app/backend/utils.py
@@ -267,10 +267,10 @@ def apply_lime_explanation(model, image_path, class_names, num_samples=1000):
             mean = np.array([0.485, 0.456, 0.406])
             std = np.array([0.229, 0.224, 0.225])
             img = (img - mean) / std
-            img_tensor = torch.from_numpy(img).permute(2, 0, 1)
+            img_tensor = torch.from_numpy(img).permute(2, 0, 1).float()
             batch.append(img_tensor)
 
-        batch_tensor = torch.stack(batch)
+        batch_tensor = torch.stack(batch).float()  # Ensure float32
 
         with torch.no_grad():
             outputs = model(batch_tensor)
